requirements.txt
asgiref==3.7.2
Django==4.2.7
django-jazzmin==2.6.0
django-resized==1.0.2
Pillow==10.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
beautifulsoup4==4.12.2
requests==2.31.0
